src/requirements.txt
# requirements.txt
azure-identity>=1.12.0
azure-ai-ml>=1.4.0
azure-mgmt-resource>=21.1.0
pandas>=1.3.5
numpy>=1.21.0
scikit-learn>=1.0.2
matplotlib>=3.5.1
requests>=2.28.0
mlflow>=2.0.0
joblib>=1.1.0
